Track limits now with correct direction, fully documented with Doxygen, treeSearch encapsulated.

+++ src/utils/Params.cpp
@@ -1,3 +1,13 @@
+/**
+ * @file Param.cpp
+ * @author Oriol Gorriz (origovi2000@gmail.com)
+ * @brief Contains the Param class member functions implementation
+ * @version 1.0
+ * @date 2022-10-31
+ * 
+ * @copyright Copyright (c) 2022 BCN eMotorsport
+ */
+
 #include "utils/Params.hpp"
 
 /* ----------------------------- Private Methods ---------------------------- */
@@ -23,7 +33,7 @@ Params::Params(ros::NodeHandle *const nh) {
   nh->param<double>(ns + "/max_angle_diff", wayComputer.max_angle_diff, 0.6);
   nh->param<double>(ns + "/edge_len_diff_factor", wayComputer.edge_len_diff_factor, 0.5);
   nh->param<int>(ns + "/max_search_options", wayComputer.max_search_options, 2);
-  nh->param<float>(ns + "/max_next_heuristic", wayComputer.max_next_heuristic, 3.0);
+  nh->param<double>(ns + "/max_next_heuristic", wayComputer.max_next_heuristic, 3.0);
   nh->param<float>(ns + "/heur_dist_ponderation", wayComputer.heur_dist_ponderation, 0.6);
   nh->param<float>(ns + "/max_treeSearch_time", wayComputer.max_treeSearch_time, 0.05);
   // Way
@@ -36,7 +46,4 @@ Params::Params(ros::NodeHandle *const nh) {
   nh->param<std::string>(ns + "/marker_topics/triangulation", visualization.triangulation_topic, "/AS/P/urinay/markers/triangulation");
   nh->param<std::string>(ns + "/marker_topics/midpoints", visualization.midpoints_topic, "/AS/P/urinay/markers/midpoints");
   nh->param<std::string>(ns + "/marker_topics/way", visualization.way_topic, "/AS/P/urinay/markers/way");
-  // nh->param<std::string>(ns + "/marker_topics/markers_full_topic", visualization.markers_full_topic, "/AS/P/urinay/markers/full");
-  // nh->param<std::string>(ns + "/marker_topics/markers_partial_topic", visualization.markers_partial_topic, "/AS/P/urinay/markers/partial");
-
 }
